Move suggest plan to its own module
