+++ nidx/src/searcher.rs
@@ -18,6 +18,7 @@ mod index_cache;
 mod query_language;
 mod plan {
     pub mod prefilter;
+    pub mod suggest;
 }
 mod query_planner;
 mod shard_merge;
